Formatted code according to PEP

+++ src/remote_sensing_processor/common/rasterize.py
@@ -22,24 +22,24 @@ def rasterize_vector(vector, raster, burn_value, output_file, nodata):
     if len(shape) > 2:
         shape = shape[-2:]
     rasterized = features.rasterize(geom_value,
-                                    out_shape = shape,
-                                    transform = transform,
-                                    fill = nodata,
-                                    dtype = shapes[burn_value].dtype)
+                                    out_shape=shape,
+                                    transform=transform,
+                                    fill=nodata,
+                                    dtype=shapes[burn_value].dtype)
     # Write
     with rio.open(
         output_file,
         'w',
-        driver = 'GTiff',
-        height = img.shape[1],
-        width = img.shape[2],
-        count = img.shape[0],
-        dtype = shapes[burn_value].dtype,
-        compress = 'deflate',
-        PREDICTOR = 2,
-        ZLEVEL = 9,
-        crs = crs,
-        transform = transform,
-        nodata = nodata
+        driver='GTiff',
+        height=img.shape[1],
+        width=img.shape[2],
+        count=img.shape[0],
+        dtype=shapes[burn_value].dtype,
+        compress='deflate',
+        PREDICTOR=2,
+        ZLEVEL=9,
+        crs=crs,
+        transform=transform,
+        nodata=nodata
     ) as outfile:
         outfile.write(rasterized, 1)
